requirements-dev.txt
pre-commit
pre-commit-hooks
black
isort
autoflake
django-stubs
